+++ CHANGELOG.md
@@ -1,1 +1,3 @@
-* Fixed Duplicate Transaction detection not using absolute value of amount
+* Report now has separate "sum expenses" and "net expenses" columns when an "income" column exists. 
+* Fixed report's Export-to-CSV not including the averages when toggled to show averages. 
+* Fixed Yearly report labeling the leftmost column "Month" instead of "Year"